--- a/app/controllers/api/v1/spaces_controller.rb
+++ b/app/controllers/api/v1/spaces_controller.rb
@@ -5,6 +5,8 @@
 if defined?(Api::V1::ApplicationController)
   class Api::V1::SpacesController < Api::V1::ApplicationController
     account_load_and_authorize_resource :space, through: :team, through_association: :spaces
+    
+    before_action :preload_associations, only: [:index, :show]
 
     # GET /api/v1/teams/:team_id/spaces
     def index
@@ -38,6 +40,21 @@ if defined?(Api::V1::ApplicationController)
     end
 
     private
+
+    def preload_associations
+      # Optimize queries by eager loading commonly accessed associations
+      if action_name == 'index' && @spaces
+        @spaces = @spaces.includes(:team, :access_passes, experiences: [:streams, :access_grants])
+      elsif action_name == 'show' && @space
+        # For single space, load comprehensive associations for detailed view
+        @space = Space.includes(
+          :team, 
+          :access_passes, 
+          experiences: [:streams, :access_grants], 
+          access_grants: [:user]
+        ).find(@space.id)
+      end
+    end
 
     module StrongParameters
       # Only allow a list of trusted parameters through.