--- a/app/controllers/public/spaces_controller.rb
+++ b/app/controllers/public/spaces_controller.rb
@@ -6,16 +6,18 @@
 
 class Public::SpacesController < Public::ApplicationController
   def show
-    # Direct slug lookup since we're using root-level routes
-    @space = Space.friendly.find(params[:space_slug])
+    # Direct slug lookup with eager loading to prevent N+1 queries
+    @space = Space.friendly
+      .includes(:team, experiences: [:streams, :access_grants], :access_passes, :access_grants)
+      .find(params[:space_slug])
 
     # Ensure the space is published for public viewing
     unless @space.published?
       raise ActiveRecord::RecordNotFound, "Space not found or not published"
     end
 
-    # Load related data for public display
-    @experiences = @space.experiences if @space.respond_to?(:experiences)
+    # Load related data for public display (already eager loaded)
+    @experiences = @space.experiences.published if @space.respond_to?(:experiences)
     @total_members = begin
       @space.total_members
     rescue
@@ -24,6 +26,9 @@ class Public::SpacesController < Public::ApplicationController
   end
 
   def index
-    @spaces = Space.published.includes(:team, :experiences)
+    # Eager load all necessary associations to prevent N+1 queries
+    @spaces = Space.published
+      .includes(:team, :access_passes, experiences: [:streams], access_grants: [:user])
+      .order(:name)
   end
 end