--- a/app/models/experience.rb
+++ b/app/models/experience.rb
@@ -4,11 +4,11 @@ class Experience < ApplicationRecord
 
   # 🚅 add attribute accessors above.
 
-  belongs_to :space
+  belongs_to :space, counter_cache: true
   # 🚅 add belongs_to associations above.
 
   has_many :access_grants, as: :purchasable, dependent: :destroy
-  has_many :streams, dependent: :destroy
+  has_many :streams, dependent: :destroy, counter_cache: true
   has_many :access_pass_experiences, dependent: :destroy
   has_many :access_passes, through: :access_pass_experiences
   # 🚅 add has_many associations above.