## Summary
Implements a complete performance optimization strategy for Backstage Pass, addressing Issue #20 with comprehensive database, caching, and monitoring improvements.

## Database Optimizations
- ✅ Performance indexes for common query patterns (spaces, experiences, streams, access_grants, etc.)
- ✅ Counter caches to eliminate expensive COUNT queries  
- ✅ Zero-downtime index creation with `algorithm: :concurrently`
- ✅ Comprehensive foreign key indexing

## Caching Strategy  
- ✅ Multi-environment cache configuration (Redis/file/memory stores)
- ✅ Model-level caching for expensive operations (`total_viewers`, `total_members`, `role_for_user`)
- ✅ Automatic cache invalidation on data changes
- ✅ Cache key generation helpers and versioning

## N+1 Query Prevention
- ✅ Eager loading in public and API controllers
- ✅ Optimized Space and Experience controllers with `preload_associations`
- ✅ Comprehensive association loading patterns

## Performance Monitoring & Analysis
- ✅ Real-time monitoring for slow queries (>100ms dev, >50ms prod)
- ✅ N+1 query detection in development environment
- ✅ Slow request tracking (>500ms prod, >1000ms dev)  
- ✅ Memory usage monitoring and cache hit rate tracking
- ✅ Performance analysis rake tasks with benchmarking

## Key Files Changed
- **Models**: Added caching to Space, AccessGrant models with invalidation callbacks
- **Controllers**: Optimized Public::Spaces, API::V1::Spaces, API::V1::Experiences with eager loading
- **Migrations**: Added performance indexes and counter cache columns with backfill
- **Initializers**: New caching and performance monitoring configurations
- **Rake Tasks**: Comprehensive performance analysis and benchmarking tools

## Performance Impact
- **Database**: Faster queries through strategic indexing
- **Memory**: Reduced database load through counter caches
- **Response Times**: Faster page loads through caching and N+1 prevention
- **Monitoring**: Real-time visibility into performance bottlenecks

## Test Plan
- [x] Migrations run successfully with zero-downtime index creation
- [x] Counter cache backfill completes without errors
- [x] StandardRB linting passes
- [x] All performance monitoring initializes correctly
- [ ] Performance benchmarks show improvement over baseline
- [ ] Cache invalidation works correctly on data changes

🤖 Generated with [Claude Code](https://claude.ai/code)

## Summary by Sourcery

Implement a comprehensive performance optimization suite by adding strategic database indexes and counter caches, introducing model-level caching with automatic invalidation, preventing N+1 queries via eager loading, and integrating real-time performance monitoring and analysis tools.

New Features:
- Configure multi-environment cache stores with versioned cache key helpers
- Add real-time performance monitoring for slow queries, N+1 detection, slow requests, memory usage, and cache hit rates
- Provide performance Rake tasks for analysis, benchmarking, cache clearing, counter cache resets, and reporting
- Introduce eager loading in public and API controllers to prevent N+1 queries

Enhancements:
- Add counter caches across Space, Experience, Stream, AccessPass, AccessGrant, and Team models with backfill support
- Create zero-downtime concurrent indexes on common query columns
- Implement model-level caching for expensive methods (total_viewers, total_members, role_for_user) with automatic invalidation
- Configure initializers for caching and performance monitoring across environments